[hls-examples] working example for VecincrXcel
I had to use ap_wait() statements to enforce ordering between multiple
stream operations. Will check with Zhiru to understand the semantics and on
how the scheduling actually works. Apart, the test memory certianly has a
bug. For this version of the code I hacked the test harness to confirm the
same and did not have to make any changes to the generated verilog.