[dstu] Test harness for HLS-based ITU (details)
Initial commit for the HLS generated verilog files, the wrapper, and the
test harness for the ITU. Currently, all tests for vectors of primitive
data-types pass with memory delays! NOTE: The test harness still is hacky
as it ties the memory response ready to 1 based on a potential bug in the
test memory model.