[itu_ptrs] BIG COMMIT: PolyDSU composition
This commit has a PolyDSU design that is composed of the configuration
logic, configuration state, polydsu router, vector and list polydsus,
polydsu funnel, and the mem-funnel router. I added a new unit for vectors,
updated the test harness, and a couple of other misc helper units. I also
fixed a minor bug in the list rtl unit wrt to the handshaking protocol. All
the unit tests pass.