[xcel/ex_sort] Initial commit for the xcel wrapper
Implemented an xcel wrapper class that is templated based on the number of
xcel registers. The class provides functions to configure xcel registers,
signal done with a user-specified value, and read xcel registers. This
should make creating HLS accelerators easier. However, there are two key
issues that need to be addressed:

  - Packing xcelreq/xcelresp channels: Currently, the tool runs into
    the issue where it cannot pack the accelerator request/response
    channels. As a work around I declare them as global variables and
    declare them as extern variables in the XcelWrapper file. This
    definitely needs to be fixed!!!!
  - Ran into bugs that involve synchronizing before and after the function
    call the xcel function. Current work around was to inline the sort
    function but we need a way to split the scheduling of two different
    transactions in seperate states which may or may not have dependences.