[mem,ex_sort] mem msgs parameterized based on data/len/opq bits
The memory messages are not parameterized based on the data_nbits,
len_nbits, and the opq_nbits. NOTE: I did not parameterize the memory
messages based on the addr bits as we are going to work with only a 32-bit
architecture for this project.

TBD: Change TestMem to work for wider memory message types.